pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financial-tools-mcp"
version = "0.1.0"
description = "Financial tools including MCP exchange rate service and invoice data extraction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "requests==2.32.3",
    "python-dateutil==2.9.0",
    "mcp>=1.9.2",
    "claude-code>=0.0.1",
    "azure-ai-documentintelligence>=1.0.2",
    "python-dotenv>=1.1.1",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "requests-mock>=1.12.1",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest",
    "requests-mock"
]

[tool.black]
line-length = 140

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = [
    "E501",  # handled by line-length
    "E203",  # whitespace before ':'
]
